add quickstart for a postgres agent with mcp
