--- quickstarts/08-data-agent-chainlit/users.sql
@@ -1,13 +1,0 @@
-INSERT INTO users (name, email, password_hash)
-VALUES
-  ('Alice Smith', 'alice@example.com', 'hash1'),
-  ('Bob Johnson', 'bob@example.com', 'hash2'),
-  ('Charlie Lee', 'charlie@example.com', 'hash3'),
-  ('Diana King', 'diana@example.com', 'hash4'),
-  ('Ethan Moore', 'ethan@example.com', 'hash5'),
-  ('Fiona Davis', 'fiona@example.com', 'hash6'),
-  ('George Hall', 'george@example.com', 'hash7'),
-  ('Hannah Scott', 'hannah@example.com', 'hash8'),
-  ('Ian Wright', 'ian@example.com', 'hash9'),
-  ('Julia Baker', 'julia@example.com', 'hash10');
-
