+++ quickstarts/08-data-agent-mcp-chainlit/app.py
@@ -2,15 +2,14 @@ import chainlit as cl
 from dapr_agents import Agent
 from dapr_agents.tool.mcp.client import MCPClient
 from dotenv import load_dotenv
-import psycopg
-import os
+from get_schema import get_table_schema_as_dict
 
 load_dotenv()
 
 instructions = [
     "You are an assistant designed to translate human readable text to postgresql queries. "
-    "Your primary goal is to provide accurate SQL queries based on the user request."
-    "If something is unclear or you need more context, ask thoughtful clarifying questions. "
+    "Your primary goal is to provide accurate SQL queries based on the user request. "
+    "If something is unclear or you need more context, ask thoughtful clarifying questions."
 ]
 
 agent = {}
@@ -60,56 +59,11 @@ async def main(message: cl.Message):
         content=result,
     ).send()
 
-    result_set = await agent.run("Execute the following sql query: " + result)
+    result_set = await agent.run("Execute the following sql query and always return a table format unless instructed otherwise. If the user asks a question regarding the data, return the result and formalize an answer based on inspecting the data: " + result)
     await cl.Message(
         content=result_set,
     ).send()
 
-def get_table_schema_as_dict():
-    conn_params = {
-        "host": os.getenv("DB_HOST"),
-        "port": os.getenv("DB_PORT"),
-        "dbname": os.getenv("DB_NAME"),
-        "user": os.getenv("DB_USER"),
-        "password": os.getenv("DB_PASSWORD")
-    }
-
-    schema_data = {}
-
-    try:
-        with psycopg.connect(**conn_params) as conn:
-            with conn.cursor() as cur:
-                # Get all table names with their schemas
-                cur.execute("""
-                    SELECT table_schema, table_name
-                    FROM information_schema.tables
-                    WHERE table_type = 'BASE TABLE' AND table_schema NOT IN ('pg_catalog', 'information_schema')
-                    ORDER BY table_schema, table_name;
-                """)
-                tables = cur.fetchall()
-
-                for schema, table in tables:
-                    schema_data[f'{schema}.{table}'] = []
-                    cur.execute("""
-                        SELECT column_name, data_type, is_nullable, column_default
-                        FROM information_schema.columns
-                        WHERE table_schema = %s AND table_name = %s
-                        ORDER BY ordinal_position;
-                    """, (schema, table))
-                    columns = cur.fetchall()
-
-                    for col in columns:
-                        schema_data[f'{schema}.{table}'].append({
-                            "column_name": col[0],
-                            "data_type": col[1],
-                            "is_nullable": col[2],
-                            "column_default": col[3]
-                        })
-                        
-                    return schema_data
-
-    except Exception as e:
-        return False
 
 def create_prompt_for_llm(schema_data, user_question):
     prompt = "Here is the schema for the tables in the database:\n\n"
