+++ quickstarts/08-data-agent-mcp-chainlit/components/conversationmemory.yaml
