+++ quickstarts/08-data-agent-mcp-chainlit/requirements.txt
@@ -1,4 +1,4 @@
-dapr-agents>=0.5.0
+dapr-agents>=0.5.1
 python-dotenv
 chainlit
 psycopg
